feat: vector-precision experiment and float16 storage default (PR 9) (#35)
* feat: vector-precision experiment and float16 storage default (PR 9)

Add the retrieval-precision experiment the remediation plan gates vector
reduction on: a deterministic judged corpus, a shared Kendall tau-b top-k
rank metric, and a benchmark that compares float32/float16 storage under
exact and SQ8-quantized HNSW search against a float32-exact numpy
reference, reporting recall, rank correlation, hybrid latency, build time,
and physical bytes per variant (CLI: benchmark precision).

Measured with the real model at 40 and 20,000 passages: float16 storage
with exact search shows no recall or rank loss and roughly halves vector
bytes, so chunk vectors now default to float16 behind schema version 5.
CODE_INDEXING_VECTOR_STORAGE=float32 restores the previous layout; a dtype
flip in either direction marks existing partitions rebuild_required via the
partition's own on-disk vector column. HNSW-SQ8 failed the gates and stays
opt-in.

* fix: review fixes for precision benchmark validation and chunk schema dtype

- Reject --passages below the retrieval-corpus topic minimum with the
  same INVALID_CONFIGURATION error as other bad CLI values, instead of
  a bare ValueError traceback from build_retrieval_corpus
- Make vector_dtype a required parameter of chunk_arrow_schema /
  _chunk_schema so no call site silently relies on the float16 default
- Document that incompatibility_reason's dtype check reads the
  partition's on-disk schema while the other checks are registry-only

+++ tests/test_memory_acceptance.py
@@ -12,6 +12,7 @@ import struct
 import sys
 from pathlib import Path
 
+import pyarrow as pa
 import pytest
 from benchmark_index_memory import (
     MIB,
@@ -240,7 +241,7 @@ def test_commit_batches_stay_bounded_far_below_the_memory_ceiling(tmp_path: Path
         tmp_path / "staging",
         "project-1",
         file_schema=LanceStore.file_arrow_schema(),
-        chunk_schema=LanceStore.chunk_arrow_schema(4),
+        chunk_schema=LanceStore.chunk_arrow_schema(4, pa.float16()),
         reference_schema=LanceStore.reference_arrow_schema(),
     )
     job.begin()
